Refactor community components
