+++ frontend/src/features/participant/community/CommunityPage.jsx
@@ -1,0 +1,119 @@
+import { useState } from 'react';
+import Diversity3OutlinedIcon from '@mui/icons-material/Diversity3Outlined';
+import LocalFloristOutlinedIcon from '@mui/icons-material/LocalFloristOutlined';
+import MenuBookOutlinedIcon from '@mui/icons-material/MenuBookOutlined';
+import {
+  communityPosts,
+  communityResources,
+  supportCircles,
+} from './communityMockData';
+import {
+  COMMUNITY_GUIDELINES_VERSION,
+  getAcceptedGuidelinesVersion,
+  saveAcceptedGuidelinesVersion,
+} from './communityGuidelinesStorage';
+import BirthdayCard from './components/BirthdayCard';
+import CommunityGuidelinesCard from './components/CommunityGuidelinesCard';
+import CommunityGuidelinesModal from './components/CommunityGuidelinesModal';
+import CommunityPostCard from './components/CommunityPostCard';
+import CommunityStreakCard from './components/CommunityStreakCard';
+import CreatePostCard from './components/CreatePostCard';
+
+export default function CommunityPage() {
+  const [showGuidelinesModal, setShowGuidelinesModal] = useState(
+    () => getAcceptedGuidelinesVersion() !== COMMUNITY_GUIDELINES_VERSION,
+  );
+
+  const handleGuidelinesContinue = () => {
+    saveAcceptedGuidelinesVersion();
+    setShowGuidelinesModal(false);
+  };
+
+  return (
+    <section className="community-page" aria-labelledby="community-page-title">
+      {showGuidelinesModal && <CommunityGuidelinesModal onContinue={handleGuidelinesContinue} />}
+
+      <header className="community-page__header">
+        <div>
+          <span>Participant community</span>
+          <h1 id="community-page-title">Community</h1>
+          <p>Connect, share, and support each other in a safe space.</p>
+        </div>
+        <span className="community-page__header-icon" aria-hidden="true">
+          <Diversity3OutlinedIcon />
+        </span>
+      </header>
+
+      <div className="community-page__layout">
+        <main className="community-page__feed" aria-label="Community feed">
+          <CreatePostCard />
+
+          <section className="community-page-card community-page-card--intro">
+            <span className="community-page-card__icon">
+              <LocalFloristOutlinedIcon />
+            </span>
+            <div>
+              <h2>Today in the community</h2>
+              <p>Stories, reflections, and encouragement from participants walking a similar path.</p>
+            </div>
+          </section>
+
+          {communityPosts.map((post) => (
+            <CommunityPostCard post={post} key={`${post.author}-${post.title}`} />
+          ))}
+        </main>
+
+        <aside className="community-page__rail" aria-label="Community sidebar">
+          <BirthdayCard />
+          <CommunityStreakCard />
+          <CommunityGuidelinesCard />
+
+          <section className="community-page-card">
+            <div className="community-page-card__heading">
+              <span className="community-page-card__icon">
+                <Diversity3OutlinedIcon />
+              </span>
+              <div>
+                <span>Circles</span>
+                <h2>Support Spaces</h2>
+              </div>
+            </div>
+            <div className="community-circle-list">
+              {supportCircles.map((circle) => {
+                const Icon = circle.icon;
+                return (
+                  <article className="community-circle-item" key={circle.title}>
+                    <span>
+                      <Icon fontSize="small" />
+                    </span>
+                    <div>
+                      <strong>{circle.title}</strong>
+                      <small>{circle.meta}</small>
+                    </div>
+                  </article>
+                );
+              })}
+            </div>
+          </section>
+
+          <section className="community-page-card community-page-card--soft">
+            <div className="community-page-card__heading">
+              <span className="community-page-card__icon">
+                <MenuBookOutlinedIcon />
+              </span>
+              <div>
+                <span>Shared care</span>
+                <h2>Community Resources</h2>
+              </div>
+            </div>
+            <ul className="community-resource-list">
+              {communityResources.map((resource) => (
+                <li key={resource}>{resource}</li>
+              ))}
+            </ul>
+          </section>
+        </aside>
+      </div>
+    </section>
+  );
+}
