+++ frontend/src/features/participant/community/components/CommunityPostCard.jsx
@@ -1,0 +1,41 @@
+import ChatBubbleOutlineOutlinedIcon from '@mui/icons-material/ChatBubbleOutlineOutlined';
+import FavoriteBorderOutlinedIcon from '@mui/icons-material/FavoriteBorderOutlined';
+import VolunteerActivismOutlinedIcon from '@mui/icons-material/VolunteerActivismOutlined';
+import CommentsPreview from './CommentsPreview';
+
+export default function CommunityPostCard({ post }) {
+  return (
+    <article className={`community-page-post community-page-post--${post.tone}`}>
+      <header>
+        <span className="community-page-post__avatar">{post.initials}</span>
+        <div>
+          <strong>{post.author}</strong>
+          <small>{post.time}</small>
+        </div>
+        <span className="community-page-post__topic">{post.topic}</span>
+      </header>
+      <div className="community-page-post__content">
+        <h3>{post.title}</h3>
+        <p>{post.body}</p>
+      </div>
+      <footer>
+        <button type="button">
+          <FavoriteBorderOutlinedIcon fontSize="small" />
+          Like
+          <span>{post.likes}</span>
+        </button>
+        <button type="button">
+          <ChatBubbleOutlineOutlinedIcon fontSize="small" />
+          Comment
+          <span>{post.comments}</span>
+        </button>
+        <button type="button">
+          <VolunteerActivismOutlinedIcon fontSize="small" />
+          Support
+          <span>{post.support}</span>
+        </button>
+      </footer>
+      <CommentsPreview comments={post.previewComments} />
+    </article>
+  );
+}
